get dependencies from composer
