+++ common/ext/class.GenerisInstaller.php
@@ -24,6 +24,7 @@
 use oat\generis\persistence\PersistenceManager;
 use oat\generis\model\data\Ontology;
 use oat\generis\persistence\sql\SchemaProviderInterface;
+use oat\oatbox\extension\exception\ManifestNotFoundException;
 
 /**
  * Custom extension installer for generis
@@ -41,7 +42,7 @@ class common_ext_GenerisInstaller extends common_ext_ExtensionInstaller
      * @throws common_Exception
      * @throws common_ext_ExtensionException
      * @throws common_ext_InstallationException
-     * @throws common_ext_ManifestNotFoundException
+     * @throws ManifestNotFoundException
      */
     public function install()
     {
