+++ common/oatbox/extension/exception/ManifestNotFoundException.php
@@ -19,20 +19,21 @@
  *
  */
 
+namespace oat\oatbox\extension\exception;
+
 /**
- * An exception that occurs in the context of Extension Manifests.
+ * This exception must be thrown when an Extension Manifest is requested but not found
  *
  * @access public
  * @author Jerome Bogaerts, <jerome.bogaerts@tudor.lu>
  * @package generis
 
  */
-class common_ext_ManifestException extends common_ext_ExtensionException
+class ManifestNotFoundException extends ManifestException
 {
-    // --- ASSOCIATIONS ---
 
-
-    // --- ATTRIBUTES ---
-
-    // --- OPERATIONS ---
+    public function getSeverity()
+    {
+        return \common_Logger::WARNING_LEVEL;
+    }
 }
